backend/requirements.txt
Flask
Flask-CORS
yfinance
pandas
numpy
nltk
google-generativeai
python-dotenv
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
joblib>=1.3.0
keras>=3.0.0